fix : MemberAuthSupport.class 를 통해 클래스 분리
