+++ src/main/java/atwoz/atwoz/member/command/application/member/MemberAuthService.java
@@ -5,14 +5,10 @@ import atwoz.atwoz.auth.domain.TokenRepository;
 import atwoz.atwoz.common.enums.Role;
 import atwoz.atwoz.member.command.application.member.dto.MemberLoginServiceDto;
 import atwoz.atwoz.member.command.application.member.exception.BannedMemberException;
-import atwoz.atwoz.member.command.application.member.exception.MemberLoginConflictException;
 import atwoz.atwoz.member.command.domain.member.Member;
 import atwoz.atwoz.member.command.domain.member.MemberCommandRepository;
-import jakarta.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
@@ -22,13 +18,13 @@ import java.time.Instant;
 public class MemberAuthService {
 
     private final MemberCommandRepository memberCommandRepository;
+    private final MemberAuthSupport memberAuthSupport;
     private final TokenProvider tokenProvider;
     private final TokenRepository tokenRepository;
 
     @Transactional
     public MemberLoginServiceDto login(String phoneNumber) {
         Member member = createOrFindMemberByPhoneNumber(phoneNumber);
-
         if (member.isBanned()) {
             throw new BannedMemberException();
         }
@@ -45,15 +41,6 @@ public class MemberAuthService {
     }
 
     private Member createOrFindMemberByPhoneNumber(String phoneNumber) {
-        return memberCommandRepository.findByPhoneNumber(phoneNumber).orElseGet(() -> create(phoneNumber));
-    }
-
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    protected Member create(String phoneNumber) {
-        try {
-            return memberCommandRepository.save(Member.fromPhoneNumber(phoneNumber));
-        } catch (DataIntegrityViolationException e) {
-            throw new MemberLoginConflictException(phoneNumber);
-        }
+        return memberCommandRepository.findByPhoneNumber(phoneNumber).orElseGet(() -> memberAuthSupport.create(phoneNumber));
     }
 }
